Refactor mobile app to use shared vault models (#900)
