+++ apps/mobile-app/components/credentials/details/AliasDetails.tsx
@@ -1,5 +1,5 @@
 import { IdentityHelperUtils } from '@/utils/shared/identity-generator';
-import { Credential } from '@/utils/types/Credential';
+import type { Credential } from '@/utils/shared/models/vault';
 
 import FormInputCopyToClipboard from '@/components/form/FormInputCopyToClipboard';
 import { ThemedText } from '@/components/themed/ThemedText';
